Assignment 2 | Mock Data를 활용하여 여러 개의 피드 구현
