+++ public/data/commentsData.json
@@ -1,0 +1,6 @@
+[
+  {
+    "id": "1",
+    "content": "댓글1"
+  }
+]
