+++ src/pages/gunwoong/Main/Main.js
@@ -1,26 +1,17 @@
-import React, { useState } from 'react';
-import Comment from './Comment';
+import React, { useEffect, useState } from 'react';
+import Feed from './Feed';
 import ASIDE_INFO from './mainFooterData';
 import './Main.scss';
 import '../../../styles/common.scss';
 
 const Main = () => {
-  const [comment, setComment] = useState('');
-  const [commentArray, setCommentArray] = useState([]);
+  const [feedsData, setFeedsData] = useState([]);
 
-  const onChangeInput = e => {
-    setComment(e.target.value);
-  };
-
-  //삼항연산자 사용해서 &&연산자로 바꾸기
-  const createComment = e => {
-    if (comment.trim() !== '') {
-      let newCommentArray = [...commentArray];
-      newCommentArray.push(comment);
-      setCommentArray(newCommentArray);
-      setComment('');
-    }
-  };
+  useEffect(() => {
+    fetch('/data/feedsData.json')
+      .then(res => res.json())
+      .then(data => setFeedsData(data));
+  }, []);
 
   return (
     <>
@@ -35,11 +26,7 @@ const Main = () => {
           <div className="logo-wrap__text">westagram</div>
         </div>
         <div className="search-bar-wrap nav-width">
-          <input
-            type="text"
-            placeholder="검색"
-            style={{ textAlign: 'center' }}
-          />
+          <input type="text" placeholder="검색" />
         </div>
         <div className="icons-wrap nav-width">
           <img
@@ -58,68 +45,12 @@ const Main = () => {
         </div>
       </nav>
       <div className="horizontal-line" />
-      <main>
-        <article className="feeds">
-          <div className="feeds__header">
-            <img
-              className="profile-img-gradient"
-              src="./images/gunwoong/IMG_3734.JPG"
-              alt="프로필사진"
-            />
-            <div className="feeds__header__contents">
-              <span>wecode_bootcamp</span>
-              <span>WeCode - 위코드</span>
-            </div>
-            <i className="fa-solid fa-ellipsis" />
-          </div>
-          <img
-            className="feeds__img"
-            src="./images/gunwoong/sample.png"
-            alt="피드이미지"
-          />
-          <div className="feeds__bottom-icons">
-            <i className="fa-solid fa-heart" />
-            <i className="fa-solid fa-comment" />
-            <i className="fa-solid fa-arrow-up-from-bracket" />
-            <i className="fa-solid fa-bookmark" />
-          </div>
-          <div className="feeds__likes">
-            <div className="feeds__likes-wrap">
-              <div>
-                <img
-                  className="feeds__likes-wrap__image"
-                  src="./images/gunwoong/IMG_3734.JPG
-        "
-                  alt="프로필사진"
-                />
-              </div>
-              <div className="feeds__likes-wrap__text">
-                <span>님 외 9명이 좋아합니다.</span>
-              </div>
-            </div>
-            <ul className="feeds__comment">
-              <Comment comment={comment} commentArray={commentArray} />
-            </ul>
-            <form
-              onClick={e => {
-                e.preventDefault();
-              }}
-            >
-              <input
-                className="feeds__comment__input"
-                type="text"
-                placeholder="댓글 달기..."
-                value={comment}
-                onChange={onChangeInput}
-              />
-              <button
-                className="feeds__comment__button"
-                hidden
-                onClick={createComment}
-              />
-            </form>
-          </div>
-        </article>
+      <div id="container">
+        <div className="feed-wrap">
+          {feedsData.map(feed => (
+            <Feed key={feed.id} feed={feed} />
+          ))}
+        </div>
         <aside className="main-aside">
           <div className="main-aside__header-wrap">
             <img
@@ -260,7 +191,7 @@ const Main = () => {
             ))}
           </footer>
         </aside>
-      </main>
+      </div>
     </>
   );
 };
